@@ -1,8 +1,9 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   experimental: {
-    appDir: true,
+    serverActions: true,
   },
-}
+};
 
-module.exports = nextConfig
+export default nextConfig;
+//module.exports = nextConfig;
